refactor-#164: move payment service to services
