+++ src/microservices/payment/payment.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ClientsModule } from '@nestjs/microservices';
 
-import { PaymentService } from './payment.service';
+import { createClientAsyncOptions } from '@utils/client';
+import { BookingModule } from '@booking/booking.module';
+import { BookingService } from '@booking/services';
+import { PaymentService } from './services';
 import { PaymentController } from './rest/payment.controller';
 import { PaymentMutationResolver } from './graphql/payment-mutation.resolver';
 import { PaymentQueryResolver } from './graphql/payment-query.resolver';
-import { createClientAsyncOptions } from '../../utils/client';
-import { BookingModule } from '../booking/booking.module';
-import { BookingService } from '../booking/services/booking.service';
 
 @Module({
   imports: [
