+++ src/microservices/payment/rest/payment.controller.spec.ts
@@ -1,12 +1,12 @@
 import { ClientsModule } from '@nestjs/microservices';
 import { Test, TestingModule } from '@nestjs/testing';
 
-import { PaymentService } from '../payment.service';
+import { PaymentService } from '../services';
 import { PaymentController } from './payment.controller';
-import { createClientAsyncOptions } from '../../../utils/client';
-import { UtilsModule } from '../../../utils/utils.module';
-import { BookingModule } from '../../../microservices/booking/booking.module';
-import { BookingService } from '../../../microservices/booking/services/booking.service';
+import { createClientAsyncOptions } from '@utils/client';
+import { UtilsModule } from '@utils/utils.module';
+import { BookingModule } from '../../booking/booking.module';
+import { BookingService } from '../../booking/services/booking.service';
 
 describe('PaymentController', () => {
   let controller: PaymentController;
