+++ src/microservices/payment/services/index.ts
@@ -1,0 +1,1 @@
+export * from './payment.service';
