+++ src/microservices/payment/services/payment.service.spec.ts
@@ -1,10 +1,10 @@
 import { ClientsModule } from '@nestjs/microservices';
 import { Test, TestingModule } from '@nestjs/testing';
 
-import { UtilsModule } from '../..//utils/utils.module';
-import { createClientAsyncOptions } from '../../utils/client';
-import { BookingModule } from '../booking/booking.module';
-import { BookingService } from '../booking/services/booking.service';
+import { UtilsModule } from '@utils/utils.module';
+import { createClientAsyncOptions } from '@utils/client';
+import { BookingModule } from '@booking/booking.module';
+import { BookingService } from '@booking/services/booking.service';
 import { PaymentService } from './payment.service';
 
 describe('PaymentService', () => {
