+++ src/microservices/payment/services/payment.service.ts
@@ -1,12 +1,12 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { ClientProxy } from '@nestjs/microservices';
 
-import { ClientService } from '../../utils/client';
-import { BookingService } from '../booking/services/booking.service';
-import { CreatePaymentInput } from './graphql/input';
-import { GqlPayment } from './graphql/models';
-import { CreatePaymentDto } from './rest/dto';
-import { RestPayment } from './rest/models/payment-rest';
+import { ClientService } from '@utils/client';
+import { BookingService } from '@booking/services';
+import { CreatePaymentInput } from '../graphql/input';
+import { GqlPayment } from '../graphql/models';
+import { CreatePaymentDto } from '../rest/dto';
+import { RestPayment } from '../rest/models/payment-rest';
 
 @Injectable()
 export class PaymentService {
